requirements.txt
streamlit>=1.28.0
google-genai>=0.3.0
pandas>=1.5.0